changes:  move strategies to infrastructure layer
