+++ backend/src/application/auth/use-cases/authenticate-user.ts
