+++ backend/src/infrastructure/modules/auth.module.ts
@@ -7,9 +7,9 @@ import { TokenVersionModule } from './token-version.module';
 import { AuthController } from '../controllers/auth.controller';
 import { UserPrismaRepository } from '../repositories/user.repository';
 import { AuthPrismaRepository } from '../repositories/auth.repository';
-import { JwtStrategy } from '@application/auth/strategies/jwt.strategy';
-import { LocalStrategy } from '@application/auth/strategies/local.strategy';
-import { GoogleStrategy } from '@application/auth/strategies/google.strategy';
+import { JwtStrategy } from '@/infrastructure/strategies/jwt.strategy';
+import { LocalStrategy } from '@/infrastructure/strategies/local.strategy';
+import { GoogleStrategy } from '@/infrastructure/strategies/google.strategy';
 import { AccountPrismaRepository } from '../repositories/account.repository';
 import { TokenVersionPrismaRepository } from '../repositories/token-version.repository';
 import { AccountModule } from './account.module';
