+++ backend/src/infrastructure/strategies/jwt.strategy.ts
