+++ backend/src/infrastructure/strategies/local.strategy.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { plainToInstance } from 'class-transformer';
 
-import { LoginLocalUseCase } from '../use-case';
+import { LoginLocalUseCase } from '../../application/auth/use-cases';
 import { UserProfileAccDto } from '@/infrastructure/dto/user/user-profile-acc.dto';
 import { AuthPrismaRepository } from '@/infrastructure/repositories/auth.repository';
